When search parameters are not null (record exists in db), when setting them transfer parameters only instead of overwriting entity object.

+++ src/test/java/com/artur/dualpair/server/service/user/SocialUserServiceTest.java
@@ -94,8 +94,8 @@ public class SocialUserServiceTest {
         assertNotNull(resultUser.getUpdated());
         SearchParameters searchParameters = resultUser.getSearchParameters();
         assertNotNull(searchParameters);
-        assertEquals(2, searchParameters.getMinAge());
-        assertEquals(8, searchParameters.getMaxAge());
+        assertEquals((Integer)2, searchParameters.getMinAge());
+        assertEquals((Integer)8, searchParameters.getMaxAge());
         assertTrue(searchParameters.getSearchFemale());
         assertFalse(searchParameters.getSearchMale());
     }
